py39 + refactor

+++ gym_footsteps_planning/footsteps_simulator/transform.py
@@ -8,7 +8,7 @@ def rotation(angle: float = 0) -> np.ndarray:
     # fmt: off
     return np.array([[np.cos(angle), -np.sin(angle), 0],
                       [np.sin(angle), np.cos(angle), 0],
-                      [0, 0, 1]])
+                      [0, 0, 1]], dtype=np.float32)
     # fmt: on
 
 
@@ -19,7 +19,7 @@ def translation(x: float = 0, y: float = 0) -> np.ndarray:
     # fmt: off
     return np.array([[1, 0, x],
                       [0, 1, y],
-                      [0, 0, 1]])
+                      [0, 0, 1]], dtype=np.float32)
     # fmt: on
 
 
@@ -37,8 +37,9 @@ def frame_inv(T: np.ndarray) -> np.ndarray:
     R = T[:2, :2]  # Rotation
     t = T[:2, 2:]  # Translation
     upper = np.hstack((R.T, -R.T @ t))
-    lower = np.array([0.0, 0.0, 1.0])
+    lower = np.array([0.0, 0.0, 1.0], dtype=np.float32)
     return np.vstack((upper, lower))
+
 
 def apply(T: np.ndarray, point: np.ndarray) -> np.ndarray:
     """
